Update safeCast and AddressLib

+++ src/lib/structs/Rewards.sol
@@ -4,15 +4,15 @@ pragma solidity 0.8.22;
 import {IERC20} from "cozy-safety-module-shared/interfaces/IERC20.sol";
 
 struct UserRewardsData {
-  uint128 accruedRewards;
-  uint128 indexSnapshot;
+  uint256 accruedRewards;
+  uint256 indexSnapshot;
 }
 
 struct ClaimableRewardsData {
   /// @dev The cumulative amount of claimed rewards since the last weight change. On a call to `finalizeConfigUpdates`,
   /// if the associated config update changes the rewards weights, this value is reset to 0.
   uint256 cumulativeClaimedRewards;
-  uint128 indexSnapshot;
+  uint256 indexSnapshot;
 }
 
 struct PreviewClaimableRewards {
